Player Movement
Player Controller Script added. This is attached to both the players. Different Controls for 2 players added in Input Manager and implemented in the script.
The Control Scheme is decided based on the player number public variable
The script allows the player to rotate the players on y-axis and move forward